exercicio04/requirements.txt
robotframework
robotframework-requests
jsonschema
